Fix bug where fields may be written in the wrong order
This bug in CREATE TABLE AS SELECT was introduced by 906ec7155b8757ae612b6c5684aa9b619e2c45b0
with the removal of the call to unpack(...). This of this method was to turn multi-field
channels into one channel per field, but also reordered channels to match what the operator
expected.