Fix failing unit test in java 8
The user.timezone property for the surefire plugin must match
the setting via argLine. The pom was inheriting the property from
airbase set to UTC, but argLine was set to Asia/Katmandu.

As far as I can tell, the surefire plugin executes the following steps:
- create a VM using the argLine arguments
- call setProperty for each systemPropertyVariable
- run the tests

Unlike Java 7, calling System.setProperty for user.timezone in Java 8 after the
 VM is initialized has no effect. See 0fe0a7be4d70ea49d9cf9a300c7fb284dd041787
for additional details