Minor cleanup and (potential) optimization in compiler
Currently, the code generates calls to getCursor for every channel from 0 up to the
max channel referenced in filters/projections. With this change, the code now only
calls getCursor on channels that are actually needed for evaluation.

This is a re-do of 3ea55c44bab8ef8730a95077d7890177b6b3df73 (reverted in 9eaffdd56e36f33574eda337622bfce3702f4b47)
with the bug fixed.