Add initial connector documentation
Starting documentation for built-in connectors: sys, jmx and tpch.

This is essential information for new Presto users, but it is currently
only available in a smattering of posts on the Presto Google group. This
adds some basic information into the documentation. Each section needs to
be developed, but this is a good starting point.