Update cassandra driver to 2.1.0
The 2.0.0-rc2 driver has a bug
(https://datastax-oss.atlassian.net/browse/JAVA-254) which gets
triggered when querying through presto.  An exception occurs because the
driver attempts to query a table that no longer exists, and fails.

Cassandra 2.1.0 is the current stable release, so it's probably time to
catch up.  The hector library has been removed and replaced, because it
didn't seem to work correctly with C* 2.1, and the project doesn't
appear to be very active.