Add config for metastore filtering
The metastore does not enforce canonical representation of values. For
example, a boolean column may represent the false value as "0", "false",
"False" or possibly other representations as well. This adds a config
flag "hive.assume-canonical-partition-keys" which can be set to true, if
all partition values are known to use the canonical (Java)
representation. Doing so may provide a performance increase for tables
with a large number of partitions that have non-string partition keys.