Always start the HttpRemoteTask info fetcher
Before the full async model, we had to delay starting the info fetcher
until after the task was created, or we would hammer the server because
we had no way to wait on the server until the task was created.  Now
we can start the fetcher immediately which is simpler.