Fix lost memory pool assignment requests
Workers ignore pool assignment requests, if the coordinator restarts,
until the sequence number reaches the previous sequence. This fixes it,
by including the coordinator id in the request. Also, disable the
cluster memory manager on the workers, as they would then race with the
coordinator to increment the pool assignment sequence number.