Optimize page compaction for dictionary blocks
When a page is compacted, if there are dictionary blocks with the same
source id, this property should be retained even after the page is
compacted. Group all the dictionary blocks in a page with the same
source id and compact them together. This also allows the result blocks
to use the same ids block.