Generalize type check in the client
The client currently expects an exact match for basic types (VARCHAR, etc).
In this commit, we relax the check to allow matching on the base name of a type
in preparation for upcoming changes to implement parametric VARCHAR.

Adding the change now allows us to deploy it to clients now to avoid complicated
coordinated rollouts when parametric VARCHAR is rolled out on the server side.

The upcoming changes to implement parametric VARCHAR