Optimize some TupleDomains to produce fewer expressions
Optimizations convert NOT IN queries to discontinuous ranges. For
example, NOT IN (1, 2, 3) turns into <1 || (>1 && <2) || >3. Queries
with a large number of values in the IN list become slow (or in some
cases generated classes that were too large) as a result of this.

Added an optimization to convert ranges into expressions with NOT INs
if doing so would produce fewer expressions.