Reduce extra symbols generated in predicate push down
During predicate push down the join condition may be mixed up, which
causes the push down code to consider the condition to be new, and
thus the generation of many unnecessary symbols.  These symbols are
propagated through the plan and cause extra data to be stored in hashes
and sent over the network.

This problem can be reduced by canonicalizing the old and new join
criteria expression when checking if the criteria has changed.