Update HttpRemoteTask to fetch only task status continuously
Earlier, the remote task had a ContinuousTaskInfoFetcher, which would
continuously ask each task for it's info object which could be huge. For
a coordinator that is managing hundreds of tasks the json
deserialization can be quite an overhead.

Extract the TaskStatus from the TaskInfo object and only update the
TaskStatus continuously. TaskStatus will propogate any state changes
in the task as quickly as possible. TaskInfo is updated independently
and used only for stats. The rate at which TaskInfo is updated is configurable.