Add parsing error test to distributed queries
This reproduces the issue fixed in 6ba55aaf9070c73ab05d010112f20d31ba0f367b

None of the existing tests catch it because it only propagates
to the caller if they happen during parsing (which causes the
QueryMonitor to be called inline). Errors that occur during
planning and analysis call the QueryMonitor in a background thread,
so they go to the logs, instead.