Fix `hive.orc.stream-buffer-size` property usage.
The property was unused but it seems like it was intended to be used
as default of `orc_stream_buffer_size` session property in the same
manner as the `hive.orc.max-buffer-size` is used as default for
`orc_max_buffer_size`.

This looks like typo in 12b485141bb434105a0bcad501ef620db8067aac.