Apply resources filter only to presto.env
Apply maven resource pluging filtering only to `presto.env` file
instead of to all the presto-product-tests resources.

Maven resource filtering is used to replace properties
placeholders e.g. `${project.version}` in the module resource files.
In our case we use this feature to substitute the actual presto version
in `presto.env` file which is used for running product tests on travis.
The problem is that the `presto-product-tests` module contains
many binary files with the test tables data. Resource plugin with filtering
enabled always tries to parse resources as text files and writes
the parsed result to the target even if no placeholders has been replaced.
This operation breaks the binary files.