Apply non-TupleDomain predicates to Hive partition list
When Hive connectors look up all partitions that is potentially needed,
it didn't taken advantage of the additional predicate available in
Constraint to apply the predicates that are not representable as
TupleDomain.

When Hive was migrated to the current TableLayout API, this additional
filter was not added. This is most likely caused by not thinking
thoroughly enough and trying to retain whatever the code did with the
old API.