revert "Add decimal product tests in presto-hive"
This reverts commit a279b15a04c26598f7fa0456f7db9b95492220dd. This
commit was not supposed to be merged, it was part of a larger part
backed out and edited before the merge of the DECIMAL works. This makes
the integration tests fail, due to presto-hive/src/test/sql/create-test.sql
having an 8 column table tmp_presto_test and inserting 10 columns into
it.