Rename FLOAT type to REAL
In order to better follow the SQL standard and other database implementations, we rename the 32-bit floating point type from FLOAT to REAL.
The SQL standard defines FLOAT to be a parametrized approximate numeric type, with the precision optional (see section 6.1. of ISO/IES 9075-2:2011). At the same time, it is not defined what precision should be used for FLOAT while no precision is given. The only constraints for precision provided for REAL, FLOAT and DOUBLE types are (6.1.29-31 of the same document):
precision of FLOAT is equal to or greater than the value specified (as a parameter)
precision of REAL is implementation-defined
precision of DOUBLE is greater than precision of REAL
Thus, for consistency with parameterized DECIMAL which we implemented to satisfy below constraint:
DECIMAL = DECIMAL(MAX_PRECISION)
unparametrized FLOAT should become FLOAT(MAX_PRECISION) which should be equal to DOUBLE.
That leaves REAL to be the 32-bit floating point type.