Change GroupIdNode to behave more like ProjectNode
This fixes a correctness issue in queries with multiple grouping sets
that resolve to the same set when unaliased. All inputs to
AggregationNode are now written in terms of outputs from GroupIdNode
or ProjectNode. This requires that GroupIdNode look a lot more like
ProjectNode, except that GroupIdNode just copies columns instead of
projecting them.