Revert "Make HashGenerationOptimizer insensitive to hash symbols order"
The optimization assumes that symbols used to express the preference
and those used to express the partitioning function for exchanges are
not aliased. When this is not the case, it fails to enforce a
partitioning that is consistent on both branches of the join,
resulting in an incorrect query plan.

This reverts commit 56a1db449e76710b2db837843cbe6147550d3fb1.