Remove queryBoundary from Scope
queryBoundary was designed to point to a scope at which query scope changes
(from inner to outer query). It appeared pointless. When two scopes point
at the same query (at some nesting level), both of them contain all
the query entities. So when field is accessed it is found in the first
scope.

So when a field is accessed from the other than current scope then it
must be an outer query scope in regards to the current query scope.
No need to check queryBoundary.