Fix describe with aggregations and parameters
Fixes aggregation analysis for describe input and output queries with
parameters in the output expressions.

Queries such as the following:
PREPARE my_query FROM SELECT count(*) + ? FROM nation;
DESCRIBE OUTPUT my_query;

would fail during aggregation analysis of the parameter with an
invalid parameter number error. The AggregationAnalyzer should not
attempt to analyze a supplied parameter value in the case of describe
input and output queries, as none is provided.