Fix coordinator sending infinite task DELETEs to rebooted workers
If a worker finishes rebooting before the coordinator gives up on a query,
a REMOTE_TASK_MISMATCH failure will be triggered on the coordinator. In
response, the coordinator will send a DELETE HTTP request to the worker to
abort the task. A subtle bug in the code results in the coordinator sending
such requests to the worker continuously without delay forever.