Add config option and session property to restore legacy ORDER BY behavior
We recently made a change to the column resolution rules for ORDER BY to
make them compliant with ANSI SQL. In order to ease the transition from
the old semantics, we now add a config option and session property that
controls the behavior.

The session property is "legacy_order_by". The config option is
"deprecated.legacy-order-by".