Fix NPE when Like node has no escape
This was a regression introduced by d30b70a2f1cd9b9a9e22d4b2b12ae4e714995451
and bb2cb51efed43ef68ccb0d71df5379fddb28464c.

LikePredicate.getNodes() was not handling the case where the escape is
missing (null), and it was trying to add a null to an ImmutableList.