Use WholeRowIterator in AccumuloRecordCursor
This fixes an issue where some results returned by the AccumuloRecordCursor
were incomplete.  Due to the use of a BatchScanner, the entries returned
could span multiple rows, causing incorrect results.  By using a WholeRowIterator
we receive the entire row that is then decoded and deserialized.  It
also simplifies the code!