Fix hive.config.resources property behaviour
`hive.config.resources` property allows to set additional resources
that must override the default properties for HDFS.

`Configuration#addResource` method loads the resource, and handles the
properties from that resource kind of like a defaults. So if before
some property has been set explicitly via `set*` method, that property
won't be overwritten with the value loaded from the resource.

The way how defaults are being populated to the configuration uses
`set` method. So, further properties loaded from the specified resources
can't override the defaults.

This commits loads the resource into temporary clean configuration. And
then copies all the properties to the destination using the `set` method,
so that way the defaults are being actually overwritten.