Temporarily remove type check in DomainTranslator
Since implicit coercions are switched to explicit coercions in the planner,
we expect all optimizers to maintain that invariant. The introduction of this check
demonstrates that we have optimizers that break this invriant. To revert to the existing
behavior, we should just remove this check until we can fix the optimizers.