Fix race condition in Driver interruption
When the Diver is closed there is a race between interrupting the current
driver thread and the driver thread releasing the lock. In very rare cases,
This can result in a thread being interrupted while procing another driver.

When a task finishes all Drivers are closed and any pending work interrupted.
This interruption should never result in a query failure because the driver
is only closed when the task is already complete. This change detects this
situation and fails with a new error containing the stack of the interrupting
thread.