Support coercion of non-function positions in lambda
`f(array(int), int, (*, int)->bigint)` fails during analysis for
`f(array(T), S, (S, T)->S)` because SignatureBinder fails when trying
to match bigint (in function return type) with S (which is presumed to
be int based on parameters seen so far).

An example of such an `f` is reduce function.