Remove reset method from Block interface
It is hard to reason about reset in BlockBuilder because it performs
non-appending operation.

For example, with reduce function, the result from previous invocation of
array_concat may be feed back into array_concat as input. If reset is called
while producing the next result, the input (the previous result) may no
longer be valid.