Handle create_table failures in Hive metastore more robustly
When Hive metastore responds with failure for create_table calls,
it might have succeeded. For example, first create_partition timed
out, but the table was actually added. The second call (retry) fails
with TableAlreadyExists.

This commit changes the behavior so that Presto checks with metastore
to see if the table was actually created when failure is encountered.
When it was, Presto considers the operation successful.