Add preprocessor support to CLI
When PRESTO_PREPROCESSOR environment variable is set, all commands
are piped through the specified program before being sent to the
Presto server.  The preprocessor is executed using '/bin/sh -c'.
The default exection timeout is 10 seconds and can
be overridden with PRESTO_PREPROCESSOR_TIMEOUT environment variable.
Additionally, the PRESTO_CATLOG and PRESTO_SCHEMA environment variables
are available to the preprocessor command.  If the query is not
executed in a context of a catalog or schems the variable will be
set to an empty value.