Fix check for resource group soft memory limit
Previously, queries were not allowed to run if the group was already at
the soft memory limit, which assumed that queries would need to use
memory. In particular, this prevented queries from running at all if
the limit was set to zero, which is the case for the legacy manager.