Add PruneProjectColumns rule
Migrate PruneUnreferencedOutputs handling of ProjectNode to
the iterative optimizer.  This rule finds a project that's under another
project, such that the parent doesn't use all the outputs of the child,
and drops the unused projections from the child.  Because the child is
a ProjectNode, it can trigger a subsequent rule to prune the grandchild.