Add PruneWindowColumns Rule
Migrate PruneUnreferencedOutputs handling of WindowNode to the iterative
optimizer.  This rule finds a WindowNode that's under a project, such
that the WindowNode defines a new column not needed by the parent, or
the grandchild produces columns not needed by either the parent or the
WindowNode, and remove any unneeded window functions, and creates a new
project grandchild to discard any unused columns.  Subsequent rules may
match the pattern of the new grandchild (project) over the old
grandchild, and do further pruning.