Make Util.pruneInputs return Optional<Set<Symbol>>
Make Util.pruneInputs return Optional<Set<Symbol>>, rather than
Optional<List<Symbol>>, because for most nodes we're thinking about sets
of used symbols, rather than the ordering of the output symbol list.

In two specific cases, ValueNode and TableScanNode, retain the previous
ordering explicitly, rather than losing it in the set, to avoid making
superfluous changes to the node during optimization.  The
PruneJoinColumns rule was already retaining the order of its outputs,
and this patch does not affect that behavior.