Add more complete WindowNode matching
The previous window matching pattern didn't validate all fields, and
didn't support a WindowFrame which contained Symbols (non-empty start or
end).  This fills out the WindowMatcher.Builder to allow constraints on
all the fields of WindowNode.

I was motivated to extend WindowNode matching because output pruning
needs to look at the WindowFrame symbols, and rather than extending it
in an incomplete way, I thought I'd try and do it thoroughly.  If people
like this approach, we can do it for other node types as well.