Remove incorrect effectiveness check before compacting block
The following heuristic is used in PagesIndex.compact() and
Page.compact() to decide whether a block should be compacted
to save memory:

    if (block.getSizeInBytes() < block.getRetainedSizeInBytes()) {
        ......
    }

The purpose is to avoid unnecessary compact for a block already in
compact representation. However, retained size is always larger than
the logical size, as retained size includes the block instance size.

We have optimized Block.copyRegion() so the original block will
be returned when copying a whole compact block.  Thus, this heuristic
is no longer necessary.