Change expected type to be Object instead of Block in Accumulator
The problem is WindowIndex.getObject(int,int) returns Object (as per signature) while we expect in Accumulator.addInput() the returned Object to be Block.
The expectation is correct because all Types returning Objects currently return only Blocks. Nothing else is ever returned from Type.getObject method.
Even though we always return Block there, we didn't want to lose the generality, so Window/PageIndex allows returning Object.
If we ever support non-Block objects, we'd need to add proper handling for them in AccumulatorCompiler, the same way as we do for any other type currently.

Even though the expected signature of the input function requires Block parameter, we don't need an extra checkedCast to be executed because JVM performs the runtime check to assert that the reference is of correct type.

Before this fix, Accumulator was code generated to seek for WindowIndex.getObject method returning Block which didn't exist. Now it seeks WindowIndex.getObject method returning Object.
Such an object is then passed to the @InputFunction method.