Update deprecated use of Throwables.propagate
This change updates the deprecated use of Throwables.propagate for
checked exceptions. If the type of the exception is IOException then we
throw an UncheckedIOException, for the other checked exceptions we
throw a RuntimeException.

With this change we only change the behavior for IOExceptions,
Throwables.propagate throws a RuntimeException for this case while we
now throw UncheckedIOException.