Fix query failure race in BroadcastOutputBuffer
Client buffer added to a broadcast output buffer were being allowed to be completed
cleanly. This means, if a broadcast buffer is failed before any client buffers are
added, the client stage will see a finished buffer and will initiate stage shutdown.
This results in a race between the propigation of the failure, and the client stage
shutting down.  In rare cases this can cause a query to incorrectly report a susccess
state.