Change constructor interface for VariableWidthBlockBuilder
newBlockBuilderLike() of VariableWidthBlockBuilder will set a minimal
entry count to be 64. If the block builder is going to add a single
entry only. The initial capacity can be 64 times of the entry size,
which leads to huge waste of memory. Pass in the expected bytes directly
instead of the average entry bytes.