Add environment support to db-backed resource manager
With the current db-backed resource manager to support multiple clusters
we need to have multiple databases (one database per cluster). With this change
the schema of the db-backed resource manager will support multiple clusters with
the new environment column. The db-backed resource manager will load the right
configuration based on the configured environment.