Reduce loader concurrency in BackgroundHiveSplitLoader
Previously, BackgroundHiveSplitLoader uses maxPartitionBatchSize as
the number of maximum concurrent threads per split source.
The default value for the configuration parameter is 100.
This is wasteful and unnecessary.

Before the previous rewrite of BackgroundHiveSplitLoader, the number of
threads is not directly bounded, but is finite due to other constraints.
At the time, the maximum possible number of threads possible were chosen.