Skip sending final leg of SPNEGO authentication
After authenticating the client, the server can include a response header
containing the final leg of the authentication, which the client can use
to authenticate the server. This is not needed with TLS, since the client
has already validated the identity of the server during the handshake. No
clients use this today, so it can be removed.