Add weighted-fair resource group scheduling policy
This policy allocates query slots amongst eligible resource groups in
realtime, weighted by shares. Of all competing groups, the group with
the lowest ratio of actual utilization (running queries/total queries)
to expected utilization (share/sum of shares of all eligible groups)
of the cluster will win the slot. If multiple groups have the same
ratio, a lottery is held where the number of tickets for each group is
proportional to the share.