Minor cleanup in AbstractTestBlock.assertRetainedSize
The following three classes have special handling in
AbstractTestBlock.assertRetainedSize:

- BlockEncoding
- AtomicLong
- MethodHandle

BlockEncoding and AtomicLong are used in InterleavedBlock, which has
been deprecated. MethodHandle is only used in MapBlock, and the
instances are shared among blocks created by the same MapType.

This commit removes the special handling for BlockEncoding and
AtomicLong, and add comments to explain the case for MethodHandles.