Address intermittent failures of testQueryLoggingCount
The change addresses intermittent failures of AbstractTestDistributedQueries.testQueryLoggingCount.
The failure happened because of how `beforeCompletedQueriesCount` is  determined at the beginning of test.
Code assumed that it is safe to get value of this counter as soon as all the
queries in queryManager are completed.

Yet it does not mean that stats counter was already updated, as those
two changes are not synchronized. See `SqlQueryManager.createQuery`:

Here is the code which updates stats counter for completed queries.
```
  QueryInfo info = queryExecution.getQueryInfo();
  stats.queryFinished(info);
  queryMonitor.queryCompletedEvent(info);
```

Yet according to `info.isFinalQueryInfo()` taken from
`queryManager.getAllQueryInfo()` the query is already completed.