Make SqlTaskExecution work with LocalExecutionPlan instead of Fragment
Previously, it takes in Fragment, and invokes LocalExecutionPlanner to get
the LocalExecutionPlan. However, it continues to look into some properties
in Fragment.

This commit adds the additional meta properties into LocalExecutionPlan
so that SqlTaskExecution don't look at Fragment any more (except to turn
it into LocalExecutionPlan). This commit also adds a constructor that
takes LocalExecutionPlan directly for improved testability.