Make double/real->decimal cast compatible with cast to varchar
Casting approximate numerics (DOUBLE and REAL) to DECIMAL used
encoding which exposed all digits from decimal representation of
values stored by floating point number (for digits before decimal point).

This resulted in somewhat unexpected results.

E.g.

cast (double '100000000000000000000000000000000' as decimal(38))
returned 100000000000000005366162204393472.

If you look at BigDecimal(double) constructor (which is discourage to be
used), it has the same behaviour.

After this change, the encoding of double will be compatible with
casting double to varchar, and also how it naturally is printed out in CLI.

The double '100000000000000000000000000000000' is printed as '1.0E32'
and cast to decimal(38) the value will match that.