SetBuildOperator to support yielding when memory limit is exceeded
SetBuildOperator didn't support yielding when tryRehash() asks for more
memory than what the system currently has. To support yielding, we added
the logic to handle unfinishedWork to the class, and it is done in
needsInput(), which is always called by the Driver in every iteration.
We didn't put it in getOutput() since SetBuildOperator is always the
last operator in the plan tree.

The added test would force rehashing to fail by reserving a lot of
memory, verify yielding happens correctly, then release the memory and
verify if the rehashing succeeds.