Acknowledge output buffer pages immediately when received
An exchange client round robins each output buffer and acknowledges
previously received pages. However, this pattern of fetch pages can be
slow when the fan-out is large. It is possible an output buff holds onto
a received page for a long while before the next reqest to acknowledge
the sent pages.
PR 7987 can solve this problem but it relies on HTTP/2 to be enabled.
This patch is a quick fix to mitigate the issue. It should be reverted
once PR 7987 is landed.