Fix reserved slots in ORC dictionary builder
position 0 is a reserved slot to indicate a null position in slice
dictionary writer. However, 0 is a reserved value instead of a reserved
key in the hash table. By avoiding hash collision to key = 0 is not
going to solve the problem given there is chance that value = 0. The
reason that when key = 0 indicating value = 0 is due to
DictionaryBuilder::rehash that accidentally assigns
blockPositionByHash[0] = 0. So whenever there is a hash collision to
key = 0, that position will be incorrectly treated as null. The real fix
is to not to avoid hash collision to key = 0 but to avoid any assignment
of the value of the hash table to be 0.