Use more portable expression in QueryBuilder
Not all databases understand `TRUE` and `FALSE`, so let's use more
portable expression.

In fact, this code is not currently reachable, since planner eliminates
table scan when condition is known to be false. Also, it eliminates
`TupleDomain`'s domains which contain all values. However,
`QueryBuilder` doesn't need to rely on that behavior.