Dynamic allocating ORC output buffer
ORC output buffer is used by all the output streams in the ORC writer.
However, each stream can vary in sizes due to different data types. A
buffer size should be adaptive based on data types. The patch starts
with a small buffer size and gradually grow it if necessary.