Enable operator CPU time tracking
This commit optimized cpu time tracking per operator.

OperatorTimer allows to decrease the number of the ThreadMXBean#getCurrentThreadCpuTime calls which
are considered reasonably efficient, but still not free.

Assuming that operations on the operators come one after another, we can reuse
the "endCpuTime" for the previous operation as "startCpuTime" for the next operations.