Revert check enforcing minimum number of workers during scheduling
The main goal of this check was to prevent scheduling queries when there
is a small number of active workers in the cluster to prevent those
queries from hitting local memory limits. However, it turns out that
this check can fail queries that would otherwise succeed. For example,
this check will fail DELETE queries for the Hive connector when there is
not enough active workers, but it shouldn't. Therefore, we are reverting
it until we come up with a better fix.

This reverts commit 8869ba3ccdd72e731ce4b75a1d823fbd4de8ffcc.