Cast null to output column types for INSERT
ORC writer uses ColumnarRow/Map/Array to inspect the structure of a
block. The down cast can fail if the input is of unknown type, which by
default creates fixed width blocks that cannot be decoded by either row,
map or array. This can happen when the INSERT misses the output alias.
We force a cast from null to the output column types to workaround this.