Remove maxBlockSizeInBytes in BlockBuilderStatus
Historically, we limit the size of each individual BlockBuilder since
we cannot limit the size of PageBuilder.  This is no longer the case
since 2ba32d62eac941403b50b3975eceefec4385b21e as we can limit
the PageBuilder size.

The BlockBuilder size in PageBulder is limited in an average way, i.e.
maxPageBytes / numBlocks. This can be problematic when
most BlockBuilders in a PageBuilder store primitive types,
while one BlockBuilder in the PageBuilder stores structural types.
The PageBuilder will be considered as full prematurely.