Allow pre-allocating memory for queries over threshold
If peak memory estimates are provided and the
experimental.memory-aware-execution flag is enabled, the coordinator
will not schedule a query for execution unless it can 'pre-allocate'
enough memory for the query to complete. This is currently designed
specifically for the scenario when there are a small number of
extremely memory-intensive queries and the goal is to not have them
trip over each other.