Remove documentation for log function
The function was removed in eb93fbb360b0e2e8653491b972181071a353a505.
The arguments to the function were in the wrong order according to the
SQL standard, resulting in incorrect results when queries were
translated to or from other SQL implementations. The equivalent to
`log(x, b)` is `ln(x) / ln(b)`. The function can be restored with
the deprecated.legacy-log-function config option.