Refactor Hive write path, target path and write mode decision
Historically, we use the equality of `writePath` and `targetPath`
and whether `writePath` exists to decide the write mode.
This is difficult to understand. Thus the `WriteMode` enum was
introduced in 23608e215280a9c51692fa1281a6a05b76f0152f.
However, it is only used in `SemiTransactionalMetastore`.

Also, `tagetPath` and `writePath` have to be fetched separately
from `LocationService`. And there are root path (query level)
and non-root path (partition/table level). These APIs are confusing.

This commit did the following refactors:
- Store the `WriteMode` in `LocationHandle`.
- returns a `WriteInfo` contains `targetPath`, `writePath` and the `writeMode`.