Fix incorrect pushdown of aggregation through outer join
PushAggregationThroughOuterJoin was incorrectly inferring that
a plan of the following shape produces distinct outputs on the
outer branch of the join:

- Aggregate
        group by x
   - LeftJoin
       - Project x = x
           - Aggregate
                group by x, y
       - ...

It was only testing whether the projection had identities, but it wasn't
ensuring that all inputs to the projection were present in the output.
In the example above, given:

(1, 10)
(1, 20)
(1, 20)

the pairs that come out the the aggregation will be unique, but the values
of each individual column will not. This breaks the assumption made by
that optimizer.

We fix the issue by removing projections from consideration. If there
is a no-op identity projection, it will be taken care of by
RemoveRedundantIdentityProjections.

Fixes #10592