Templatize and rename LookupSourceFactoryManager
The class was responsible for managing the data that's shared between join
build and probe. Depending on the pipeline execution strategy of build and
probe, there can be different ways data are shared between them.
Besides Hash Lookup join, this class also fits the model of nested loop join.
This commit renames it to JoinBridgeLifecycleManager.