Define all parameters in single place
It is easier when parameters are set in single place.
Also pom.xml is not the best place to set default values for parameters,
because test can called from IDE and then defaults are not set. When
parameters are set in script then it is easier to manually copy them to IDE.