Refactor HiveMetastore statistics related interfaces
- Remove optional from the Metastore statistics interfaces
- Remove column granularity from getStatistics* metastore methods

Optional.empty() and Optional.of(ImmutableMap.of()) is basically the same.
It introduces a lot of confusion of when to return an empty map, and when
to return an empty optional.

Also currently we never query statistics for a subset of columns. It is not supported
by the ConnectorMetadata interface. So this code just complicates the interfaces,
and increases the amount of memory used by the statistics cache.