Compute grouping sets during planning
This change moves the expansion of grouping sets from the analyzer
to the planner. AST nodes for GroupingElement now only describe syntax.
The analyzer only needs to process the original expressions in the
syntax tree (e.g., for the purpose of determining their type and validity)
instead of processing each instantiation of the expression in the expanded
grouping sets.

Also fixes #11120