Refactor HiveColumnStatistics
Replace maxColumnLength with maxValueSizeInBytes and averageColumnLength
with totalValuesSizeInBytes

Since what we need in the optimizer is a data size, there is no need to convert it
back and forth to averageColumnLength if it is not required by the Metastore.

The Hive Metastore stores averageColumnLength. so the totalValuesSizeInBytes gets
converted to averageColumnLength and back right before storing or reading it to/from
the Hive Metastore.