Add block/position convention to all distinct from operators
This commit adds a second implementation to distinct from operator of all
types. This 2nd implementation uses block/position convention. As a
result, for certain types, it's no longer necessary to extract the value
out of the Block in order to invoke "distinct from" operator. This is
expected to resolve performance issue previously introduced in
19230b37d352b30ef89a3552d42757d71655bfcc (which was work around with a
hack).