Implement recursion-free ATN simulator
The new implementation tracks all the possible states the ATN can be
in in a single queue instead of recursing when processing sub-rules.

This vastly simplifies the code, makes it easier to reason about and
debug. It also fixes a latent bug where some contexts were not being
visited, which missed some candidate tokens.