Fix flaky testComplexStateEstimatedSize test
This test is non deterministic, as it checks the estimated retained memory usage.
The memory usage estimate over the SliceBigArray and over the BlockBigArray is non
deterministic because of usage of the ReferenceCountMap, which stores it's values
by the non deterministic identity hash that allows collisions.

The fix runs the test 100 times and checks if 90% of runs are successful.