Fix currentConstraint when PickTableLayout resolves TableLayout
This change should have been part of 3bce2e7ced6bedcb65d75f8469d83990e77bbb6a
but was not placed there due to oversight.

As clarified in commit 3bce2e7ced6bedcb65d75f8469d83990e77bbb6a and explained
in the commit message of the same commit, currentConstraint is supposed to be
a TupleDomain that represents a predicate that every row produced by
this TableScan node is guaranteed to satisfy.

To recap, a table scan is not guaranteed to only produce rows that satisfies
the filter on top of itself. As a result, it is incorrect to have
the intersection.

This commit affects EXPLAIN IO. However, although EXPLAIN IO was introduced
before this commit, only showing successfully pushed down predicates in
EXPLAIN IO has always been the intended semantics.