Move statistics aggregations on write to system memory pool
With the change that removes the memory limit for partial aggregations the
statistics aggregations were updated to use user memory instead of system memory.

Statistics aggregations is an implicit process that the query doesn't
have much control over. It is very similar to implicit buffer allocations required
when writing ORC files. Although the amount of memory used depends on the user input
(number of columns in the table) the user may not even be aware that stats will be
collected. Therefore, it sounds like it is more suitable to use system memory
for stats collecting aggregations.

This change introduces a "useSystemMemory" parameter that sets the type of
memory that aggregation operators should use. The existing rule of using
system memory for partial aggregations is also removed. The type of memory to use is
now explicitly set in LocalExecutionPlanner based on the type of aggregation --
system memory is used for partial aggregations and for column statistics aggregations
used during writes.