Implement qdigest type and associated functions
Adds the following functions for creating and manipulating qdigests:
- `qdigest_agg`: aggregation function to create qdigest from real,
bigint or double values
- `merge`: aggregation function to merge multiple qdigests
- `value_at_quantile`: scalar function that retrieves a given
quantile from a qdigest
- `values_at_quantiles`: scalar function that retrieves a list of
quantiles from a qdigest

Weights and accuracy specifications can be set to `qdigest_agg` in the
same way they can be set on `approx_percentile`.  `qdigest`s support
casting to and from `varbinary`.