Change cost estimate formatting in PlanPrinter
- Do not print estimates if stats are unknown for all the plan nodes for the operator
- Print "?" for unknown stats if stats are known for some plan nodes for the operator

For example ScanFilterProject operation consist of 3 different plan nodes. ScanNode, FilterNode, ProjectNode.

If the cost estimate is lost, say, on Filter node, for the matter of readability it makes sense to print:

{rows: 123 (23kB), cpu: 234, memory: 345, network: 567}/{rows: ? (?), cpu: ?, memory: ?, network: ?}/{rows: ? (%s), cpu: ?, memory: ?, network: ?}

so it is clear that the estimate is lost on the filter operation.

It doesn't make much sense to always print "?" though, as it makes plan to look cumbersome without introducing any value.