Improve query plan when Hive tables have compatible bucketing
When two Hive tables have same bucketing key, but different bucket
number, the tables are considered to have compatible bucketing.
Remote exchange is not necessary to join them when the bucket numbers
are compatible.

Bucket numbers are considered compatible if the one is the multiple of
the other. In the current implementation, the multiplier is required
to be a power of two. This power-of-two constraint is not strictly
necessary and can be removed.

This change also applies when one is reading from a bucketed table
and writing into another. It also applies when group by is applied on
union all of two or more tables.