Fix query execution time calculation
Query execution begins when the query is transitioned to planning.  This was expected
by the computation that computed excution time by subtracting queued time from elapsed
time, but that is not longer correct with the addition of the resource waiting state.

The should not have a significant impact since resource waiting is a new feature that is
disabled by default.