Enable char predicates to be pushed down
When running queries with presto, if the underlying
table is of type varchar, the predicate gets pushed.

If it's of type char, it will NOT get pushed, which
causes presto to get the whole table, and then filter
locally, which is a big pain (and bad performance).

Using Presto PR #4842 as a guide, I've added support
for char types to also be passed down.  Unit tests
also updated for the type following the pattern.
Also integration tested against a mysql-like backend.