Revert lazy build hashtable
This reverts
1) commit e70f5d3089eb885a90e5d991c6626331d8ca4e97.
2) commit 714feb3f7b3d2f25215e19c6958bf484a83f935b.

PR #11791 (commit 714feb3 and e70f5d3), which lazily builds the
hashtables for maps, introduced a regression for the case where the
MapBlock is created through AbstractMapBlock.getRegion(). The hashtables
built on the MapBlock region were not updated in the original MapBlock,
thus causing hashtables repeatedly being built on the same base MapBlock.