Add option not to delete target directory on failure
When a metastore operation fails, especially due to timeout, we do
not know whether it ultimately completes. Relying on the apparent
failure to perform cleanup operations is thus unsafe.

Cleaning up new target directories can cause data loss when Presto
writes to existing partitions and replaces them in the metastore.
This occurs when the metastore call times out on the Presto side, but
eventually completes, and the metastore call to undo the operation
fails. This results in a metastore partition pointing to the
now-deleted staging directory.

This can be avoided by deferring cleanup of staging directories to an
external system that works in concert with the metastore to determine
which directories are actually referenced by partition locations.

Extracted-From: https://github.com/prestodb/presto