Change scroll behavior to zoom only when query is completed
D3 zoom did not work well when the DOM was changed, elements could
sometimes appear in incorrect positions. While the query is running
this is sometimes possible. This commit changes the behavior such that
zoom is not enabled until the DOM stops changing (i.e. the query is
complete).

Extracted-From: https://github.com/prestodb/presto