Respect original column type when inserting
In JDBC connectors, currently the insert target type for a column depends
on the Presto type chosen to represent the column's values.
This strategy doesn't support multiple original types mapped to the same
Presto type -- a single `WriteMapping` imposes a target type which not necessarily
fits them all, and so inserts may cause errors at finish time when trying to copy
the temporary table into the target table (unless the target database is able
to perform a coercion).
This is fixed by making INSERT aware of the target column types i.e. by using
the original column type in the temporary table and obtaining the write function
from the `ColumnMapping`.