Change default hive metastore refresh interval
With previous configuration, whenever someone was setting up
a ttl, refresh-interval default was used as 0, which is an invalid
configuration option for cache, which in turn has caused PrestoServer
startup failure. Changing it to Duration(Integer.MAX_VALUE, DAYS)
makes it disabled by default as well, but allows configuration
with ttl only, as refresh interval is not applied when it's greater than ttl.