Fix CorrelatedJoin decorrelation
Before this change, CorrelatedJoinNode of type RIGHT and FULL
was incorrectly rewritten to JoinNode in the case when correlated
filter was present.
Because the correlated filter was pulled into the JoinNode's filter,
and the JoinNode's type was RIGHT or FULL, all rows from the subquery
were preserved. Accordingly to the semantics of LATERAL derived relation,
only the rows which pass the correlated filter should be present
in the derived relation.
This is achieved by applying another type of JoinNode:
- INNER JoinNode for RIGHT CorrelatedJoinNode,
- LEFT JoinNode for FULL CorrelatedJoinNode.
If CorrelatedJoinNode's filter is other than ON TRUE, it is:
- applied as Projection in the case of RIGHT CorrelatedJoinNode,
- not supported in the case of FULL CorrelatedJoinNode.

All other cases of CorrelatedJoinNode with respect to join type,
join condition and presence of correlated filter, are handled
as they were before, i.e. via rewrite to JoinNode of the same type
and with filter being a conjunction of CorrelatedJoinNode's filter
and correlated filter.