Fix RemoveUnreferencedScalarSubqueries rule
Fixes incorrectly pruned unreferenced scalar input of
correlated join.
Before this change, the results were incorrect for
LEFT / FULL correlated join in the case of empty subquery.
By removing the input, empty result was obtained, while
it should be null.
Also, pruning the input was incorrect in the case of
correlation present. The remaining correlated subquery
couldn't be decorrelated.