Fix reading from large Parquet files
This is a fix for https://github.com/prestosql/presto/issues/2730. When
merging small reads, if the first range and second range are more than 2
GB apart, mergeAdjacentDiskRanges() throw sn ArithmeticException because
merging those two ranges is too big to fit in a DiskRange. The correct
behavior is to not merge those ranges because this implies the ranges
are farther apart than maxReadSizeBytes.