Support Materialized Views
This commit adds support for Materialized Views in Presto engine.
Much like a logical view, a materialized view has a SQL query associated with it.
Unlike logical view, it stores data corresponding to the SQL query.

The commit adds support for commands like CREATE MATERIALIZED VIEW, REFRESH MATERIALIZED VIEW,
SHOW CREATE MATERIALIZED VIEW and DROP MATERIALIZED VIEW.
The commit adds support for reading data from a  materialized view when it is fresh with
respect to its underlying base tables. When a materialized view is stale with respect to its
base tables, the materialized view is resolved to base tables using the associated definition.
Querying the materialized view always returns the current/fresh data irrespective of the state
of the materialized view.

A materialized view is modeled as a combination of a SQL definition and a storage table
that holds the data.

The ‘Refresh Materialized View’ command is implemented as follows:
REFRESH MATERIALIZED VIEW Implementation:
+ Refresh materialized view operation is implemented as a table writer that drops partitions from,
  deletes data from and inserts data into the storage table as needed. The source of the data is
  the query associated with the materialized view.
+ A new type of TableWriterOperator, ‘RefreshMaterializedViewTarget’ is implemented. This translates
  into two connector API calls ‘beginRefreshMaterializedView’ and ‘finishRefreshMaterializedView’.
+ StatementAnalyzer determines if the materialized view is fresh and sets the flag in Analysis.
  If the materialized view is fresh, logical planner plans the refresh operation as a no-op.
+ The ‘beginRefreshMaterializedView’ implementation for a connector is expected to do the following:
  + Start a transaction
  + Drop specified partitions of the storage table based on input parameters (applicable only
    for incremental refresh of the materialized view)
  + Delete data from specified partitions of the storage table or all of the data from the
    storage table based on input parameters (applicable for incremental refresh and full refresh
    respectively)
  + Return a ConnectorInsertTableHandle
+ The ‘finishRefreshMaterializedView’ implementation for a connector is expected to do the following:
  + Insert data into the storage table based on parameters
  + Store the table tokens for the base tables in the storage table
  + Commit the transaction.
+ Note that the refresh materialized view operation is performed in the scope of a single
  transaction in the connector.

Access control:
Given that materialized views can be seen as a combination of a view and a table,
access control for a CREATE MATERIALIZED VIEW command is a combination of access
checks for CREATE TABLE and CREATE VIEW commands.
Similarly, a REFRESH MATERIALIZED VIEW command is a combination of DELETE and INSERT
operations and access checks for this command is a combination of access checks
for DELETE and INSERT.
Lastly, a DROP MATERIALIZED VIEW access check is a combination of DROP TABLE and
DROP VIEW command.