Use source's symbol for projection instead of output symbol
In PushProjectionThroughExchange rule, there is a projection
inserted into the plan between ExchangeNode and each of it's
sources. The projection is supposed to retain source's hash symbol.
Before this change, there was added an identity assignment
on hash symbol from ExchangeNode's output. It was wrong in the
case when source used other symbol (with different name) than the
ExchangeNode.
It was fixed by using the source's symbol instead.