Use varchar for Elasticsearch passthrough query result
Presto's JSON type imposes additional constraints that are not
desirable for this use case: JSON values must be equatable and
orderable. This requires parsing and re-organizing the document
to canonicalize field ordering.

Using VARCHAR will also play better with the SQL 2016 JSON features,
which operate on binary or string data directly.