Simplify presto-benchto-benchmarks executable jar
After 1e34c653274554adddbc671e2a0dd13d10aa8bd5 the `.jar` artifact
contained classes.  This resulted in `-executable.jar` being to big and
release to maven central failing due to lack of `-javadoc.jar`. The
javadoc jar is not created because there is no Java code in the module.

This restores previous packaging, while keeping the improvement
introduced in that commit.