Do not call getTableDefaultLocation if external_location present in CTAS
In
https://github.com/prestosql/presto/commit/c402fd5c0729236195ae5a3ca91ea86fa2f53dee
CTAS was fixed to use the external_location if provided but due to the
way Optional.orElse() works it still ended up calling
getTableDefaultLocation() which will throw errors if the schema location
cannot be accessed or is missing. This can be fixed by the user
by fixing the location defined in the schema - but Hive doesn't allow
that. So the other option is to not check the location defined in the
schema until absolutely needed.

For testing a new Hive catalog is added with writes to external tables
enabled.

Fixes: #4069.