Refactor HiveMetastore API and Implementations
This commit changes the HiveMetastore API to remove methods to methods
getPartionNames() and getPartitionNamesByParts() and replace with
getPartitionNamesByFilter(). Each implementation is updated as well as
all callsites. The previous behavior is maintained as far as
translation of any non-equal Domain values to the wildcard value.
A subsequent commit will implement translation of a Domain into a
proper Glue expression for the GlueHiveMetastore GetPartitions API in
the Glue Data Catalog.