Improve Access Denied message: insufficient view owner privileges
If a user A creates a view and user B wants to access the view and
then user A loses access to the original data
then B should not be able to access data in the view.
If B gets and access denied,
they are told whether it's because they don't have access to the view
or if the view creator doesn't have access to the underlying data.