Deprecate isCharType, isVarcharType
There is no way of preventing `instanceof VarcharType` and we do no try
to discourage that. For consistency of the code base, let's standardize
on (unavoidable) `instanceof` and depart from using a method to check
that.

Note also we do not have such methods for some other parametric types,
e.g.  timestamps, decimals, arrays.