Close Postgres test resources in order
Before this change testing Postgres server was closed before the query runner.
That led to the situation where query runner was up
and running but without a dependant service.

In case of Postgres testing this was not causing any issue as query runner
was not used during that time, but this reveals a generic tests infra
issue that may affect other tests.