Iceberg connector support for materialized views
Iceberg connector allows creation, refresh and drop of materialized views.
It provides methods to determine if a given materialized view is current
with respect to underlying base table(s).
Freshness of a materialized view is determined by comparing the snapshot ID
of base table(s) at the time the materialized view was refreshed with
snapshot ID of base table(s) at the time of the check.