Add version check in file metastore
File metastore implementation is tailored for testing purposes and we
make no backward compatibility guarantees for the internal structures
used by it. However, this is not reflected in code behavior, as we make
the implementation available and selectable by the users. Prevent
accidental usage of metadata persisted with one file metastore version
and being read with a different version to avoid potential incorrect results.
Users can still opt in into previous unsafe behavior with a
configuration property.