Support multiple filter to mask rewrites for aggregation
This change enables rewriting Aggregation's filter to mask
in case when mask is already present.
It allows to add filter during optimization to an Aggregation
that had FILTER. It helps avoid dependencies between rules.

Also, after this change, unnecessary projection isn't added to the plan.
Before, every filter symbol was projected to a new mask symbol.
Now, renaming projections are skipped and a ProjectNode is only added
when there is a need to combine symbols: existing mask and filter.