Simplify assertions in Hive timestamp precision test
Since commit 58ed1c5111f942c13cc749bc35fdfcce0d19e556 we have strict
validation of timestamps returned from the engine, so the assertions
with `=`, `<` or `>` predicates are redundant.

Instead, it is more useful to precisely validate the timestamp as seen
by the engine instead of relying on the JDBC representation, and this
can be effected using `typeof` and `CAST` to `varchar`.