Remove default implementation of JDBC type mappings
This is just a refactor. `BaseJdbcClient` no longer provides default
implementation of `toPrestoType` and `toWriteMapping`. Instead, it
provides these as a separate function that a connector may or may not
call. Notably, `OracleClient` does not use these methods, because the
Oracle connector covers all its mappings explicitly. Same should be
applied to all other connectors.