Fix TopNRowNumberOperator incorrectly swapped types
TopNRowNumberOperator was previously incorrectly using the output type
order for the SimplePageWithPositionComparator strategy, when the
channels were all defined in terms of the input types. This issue is not
visible in production code because the LocalExecutionPlanner always puts
the outputs in the same order as the inputs, but this means that the
current set of tests were accidentally correct. The tests have been
updated to fail if this occurs.