Mark serialization-only methods as @Deprecated
It is not ideal to mark a method we do not plan to remove as a
`@Deprecated`, but it is still better than mark it with a hard to see
comment as "do not use".

The alternative approaches -- like introducing a new annotation just for
that purpose and add a static code analysis to detect violations are
definitely possible, but not something that can be delivered quickly.