Add Error Prone check: ThrowIfUncheckedKnownChecked
The method `Throwables.throwIfUnchecked` re-throws a given `Throwable`
if it's unchecked (`RuntimeException` or `Error`). This only makes sense
when the static type is too general to be known if the actual exception
is checked or unchecked. Wherever the static type is known to be
checked, this is a no-op and the invocation can be removed.