Avoid integer overflow when computing frame bounds
This change applies to window frame type GROUPS.
Previously, the expected group index was always
cast to int by `toIntExact`.
After this change, when the expected group index
exceeds the range of group indexes available in the
partition, all rows until the partition bound
are inclued in the frame, and the `toIntExact` call
is ommited.