Fix char predicate pushdown for MySQL, MemSQL and SQL Server
In case insensitive predicate pushdown, consider values 'A' and 'a'.
Both are equal in data source, but different in Trino.

It was safe to pushdown `x = 'a'` or `x = 'A'` as both values where
always returned and we could filter them out in Trino.

But operators like `!=`, `<` or `>` may return no value at all, which is
incorrect as Trino is case sensitive only today.

Consider also a casse of `B` and `a`. In Trino `B` is lower than `a`,
but if you compare them in remote data source which is casee
insensitive then `B` or `b` are higher `A` or `b`.

Considering the above we can only push down the predicate
for case insensitive column only when `=` operator is used.