Verify table metadata in SHOW COLUMNS
We are using information_schema which may ignore
errors when getting the list of columns for a table,
since listing columns is a requirement for some tools,
and thus failing due to a single bad table would make the system unusable.

However, when showing columns for a single table,
it is important to fail if the columns are not available,
rather than erroneously returning an empty list.
We thus ask for table metadata, which will hopefully
fail for the same reasons that would cause an empty list of
columns.

We still go through information_schema, even though we appear
to have all the needed information in the table metadata,
so that we use the same code path for all column listing.
Connectors may have different listing logic than for metadata,
and we need to perform security filtering of the returned columns.