Fix Join optimization rule
In case of FULL join having a scalar source producing no outputs,
do not replace the join with the other source if it cannot be
determined to have at least one row.

If the other source turns out to be empty, we would get empty
result after such replacement. This is incorrect. The single input row
from the scalar source should be preserved on output.