Revert "Refactor TypeCoercion.coerceTypeBase"
This reverts commit bfbe35de8953e273a5786e325aabe9720df2f439.

bfbe35de8953e273a5786e325aabe9720df2f439 was merged because it seemed to
reduce probability of a rare failure in function resolution. The failure
itself looked as if it was JIT-related. Later it was observed that even
if that commit reduced probability of such failures, it definitely did
not eliminate it. However, the problem was not observed recently, so it
is possible it stopped appearing thanks to a JDK version change
(probably from about 11.0.9 to 11.0.11 used on CI today).

The revert itself was a mechanical `git revert`.