Revert "Supports authentication using multiple ldap.user-bind-pattern value"
This reverts the following commits:
764fa5f981758503b7a89181dba79d171a2c91ed
4372719fe6636252a044116507b529df8f21ec6a
8c6c180e33aabd30ad8903cc693948b89e57b61f
2082a82ae516b9fb829d091f3098dad32fd929f4
d09e98ee43a6f9b5727f5706b91551c85d6ab8f2

The PasswordAuthenticator SPI expects an AccessDeniedException to be
thrown when authentication fails. The commits above changed the
behaviour and instead throw RuntimeException in case of failed
authentication which violates the SPI. The callers don't handle
RuntimeException which bubbles up as an internal server error instead of
a meaningful failure.

This was not caught because some tests were not run on the pull-request
(fixed in bf2b0b3e25862873b4e6100935939ae8e1adba7e).