Fix flaky TestJdbcConnection
PreparedStatement::execute may or may not block, hence the future may
end up with an exception for cancellation (which is the usual case)
or not (the flaky case).

JDBC specification does not mention whether is a blocking method or not.
It is up to the vendor. Hence, we should not depend on that it is
currently blocking most of the times.

I am not sure why it is not blocked sometimes, as this is
a very rare event.

To fix the issue:
 - lambda consumes old the results to make sure it blocks
 - assertion no longer specific failure. As it can fail in many
 different ways including: error executing query, error fetching results
 or plain interrupted exception.
 - make sure that query didn't fail or finish before cancellation