Reject Glue and Alluxio metastores from Iceberg
They are currently not supported.

Previously, Iceberg would use `HiveMetastoreModule` and thus inherited
metastore configuration from Hive connector. `IcebergMetastoreModule`
existed for validation purposes. Besides thrift and file, Iceberg
inherited 'support' for glue (which actually didn't work at runtime) and
alluxio metastores (which was not intentional).

The commit copies `HiveMetastoreModule` into `IcebergMetastoreModule`
so that the latter class defines metastore configuration for Iceberg.
The Glue metastore option is currently not supported, but will be added
in the future. The Alluxio metastore option is dropped.