Make SUPPORTS_RENAME_SCHEMA default to SUPPORTS_CREATE_SCHEMA
If a connector doesn't support creating schemas it's not possible to
test if it supports renaming schemas. So to ensure test results aren't
misleading enforce that connectors that don't support schema creation
assume that schema rename isn't supported. The assumption is explicitly
validated so where needed it can be overridden by the connector.