Fix infinite loop in optimizer
Commit 9ccfe0944e1c2b84d32a60763f30e8dee630fe0e surfaced an incompatibility between
optimizer rules that causes ping-ponging and an infinite loop when the rules are
placed in the same IterativeOptimizer.

For this query:

   SELECT orderkey FROM orders WHERE custkey IS NULL OR custkey = 370 AND orderkey = 1

The SimplifyExpressions optimizer and RemoveRedundantTableScanPredicate ping pong between:

   ((("custkey" = BIGINT '370') OR ("custkey" IS NULL)) AND (("custkey" IS NULL) OR ("orderkey" = BIGINT '1')))

and

   ((("custkey" IS NULL) OR ("custkey" = BIGINT '370')) AND (("custkey" IS NULL) OR ("orderkey" = BIGINT '1')))

This is because PushPredicateIntoTableScan.createResultingPredicate() and ExtractCommonPredicatesExpressionRewriter.extractCommonPredicates()
produce different canonicalizations of the expression. A simple, although not ideal, fix is to process the result of createResultingPredicate()
through the same rewriter used by SimplifyExpressions to use the same canonical form.