Enable Error Prone check: FormatStringAnnotation
This forces us to annotate all the places where we wrap or delegate
logs. There's one suppression, where we intentionally add one more
placeholder to the format string and Error Prone won't take it in any
form.

It is ERROR by default. Will be removed once we enable all checks.