Flip exception parameter in logs to its righful place
In some logging frameworks the exception to print is expected to be the
last format string argument, even though there's no placeholder for it.
In Airlift, though, this is different, and there's an explicit exception
parameter before the format string. Fortunately Error Prone can flag
this, although somewhat by accident (by seeing that there's no
placeholder for the exception parameter).