Pipeline buffering and upload in S3 multi-upload.
With code restructuring we needed to add separate multipartUploadStarted
flag which is both written and read from main thread (not an upload)
thread. With `waitForPreviousUploadFinish` moved below we lost
memory-barrier between main and upload thread and we could no longer use
`uploadId` to determine if multipart upload was already started.

Co-authored-by: Łukasz Osipiuk <lukasz@osipuk.net>