Fix failure when selecting a constant and all fields from row
In a query
`SELECT 1, ROW(2, 3).*`,
on the planning phase the expression `Row(2, 3).*` is unfolded
into subscript expressions: `ROW(2, 3)[1]` and `ROW(2, 3)[2]`.
Because the query also contains the literal `1`, the subscript
index in `ROW(2, 3)[1]` was rewritten into a symbol by the
`TranslationMap`. The resulting expression was `ROW(2, 3)[s]`,
which was incorrect, since the row subscript is required to be
a constant.

This change modifies the `TranslationMap`s rewrite method
so that the subscript index is not rewritten.