Implement renameSchema for ClickHouse
The current implementation executes `ALTER SCHEMA ... RENAME` unconditionally,
and let ClickHouse throw the proper errors.
The reasoning for this is that we already do this for some other methods
in some connectors. And it's possible that our assumptions today will
become false in future if there are more changes in ClickHouse.