Use ToxiProxy to forward traffic from host to Kudu tablets
Kudu tablets needs to know the host/mapped port it will be bound to in
order to configure --rpc_advertised_addresses. However when using
non-fixed ports in testcontainers, we only know the mapped port after
the container starts up.

In order to workaround this, create a proxy to forward traffic from the
host to the underlying tablets. Since the ToxiProxy container starts up
*before* kudu, we know the mapped port when configuring the kudu
tablets.