Migrate Oracle to use BaseCaseInsensitiveMappingTest
This also changes some tests to not use the session schema for testing
mapping rules because `getSession().getSchema()` returns the schema name
as set in the session while the mapping rules require the schema name to
match what actually exists in the remote database.

In case of Oracle the default session schema is not quoted during
creation in `TestingOracleServer#createConfigureScript` which causes it
to be created in uppercase. The tests create the mapping rules with the
lowercase name which causes failures.

A possible solution is to inspect
`DatabaseMetaData#storesUpperCaseIdentifiers` and adjust the value
passed to the mapping rule accordingly but it's cumbersome to access
`DatabaseMetaData` from the test so instead we opt to run these tests in
new schemas which would get created with correct case.