Do not display maven version
Remove obsolete hotfix

If maven is not downloaded for maven-wrapper,
`./mvnw` downloads it with `org.apache.maven.wrapper.Installer`.
`org.apache.maven.wrapper.Installer` in version `0.1.2-SNAPSHOT` which
we had some day before (upgraded in
0940a94a533f9e6c3daa19703c045394b75c40f2) logs to the standard output.
The newest version of maven-wrapper.jar is `3.1.0` and it logs to log
https://mvnrepository.com/artifact/org.apache.maven.wrapper/maven-wrapper

To overcome this issue `./mvnw -v` construction was used to
skip output before any actual `./mvnw` invocation which returned
valuable text output.