Reduce lock contention in FutureStateChange
Avoids unnecessary listener removal and the associated object
monitor synchronization when listeners are completed as part of a
normal FutureStateChange notification event. Only listeners that
were canceled before being notified need to be removed from the
current listener set.

Also changes the logic for sending a change notification to copy
the current listeners set into a List instead of a Set, which has
more overhead and was not necessary.