Change the type of UNSUPPORTED_TABLE_TYPE to EXTERNAL
`UNSUPPORTED_TABLE_TYPE` error code was initially introduced to
provide a more generic approach than
`HiveErrorCode.HIVE_UNSUPPORTED_FORMAT` to dealing with
the situation of selecting a table of an unsupported type
from a shared metastore (e.g. : Hive, Glue).

The internal logic from `HiveMetadata.streamTableColumns` however
made use of the error type of `HiveErrorCode.HIVE_UNSUPPORTED_FORMAT`
which is `EXTERNAL`. Due to this reason, the error type of
`UNSUPPORTED_TABLE_TYPE` will be changed as well to `EXTERNAL`.