Upgrade Error Prone to 2.12.0
`DoubleBraceInitialization` got upgraded to ERROR by default, so we can
remove it from the list of checks upgraded explicitly.

There is a bunch of spurious warnings about unread fields, which are
used in tests in a declarative way. Also, two comparison of `Boolean`
values using identity comparison, which should be OK (`Boolean` is
always interned), but let's make unboxing explicit.