Simplify version logic in plugin/SPI compatibility check
The check relied on the fact that plugin version and SPI version match
for all plugins maintained within Trino. While this is true, loading
plugin version from jar's manifest isn't exactly same as loading the
version from a maven-filtered resource. This in turn lead to
discrepancies in reported version when a tagged version was rebuilt with
local modifications. (Of course, this remains not a recommended
practice).

This commit replaces the jar's manifest and with explicit maven-filtered
resource on the plugin side. The filtered resource is within
trino-plugin-toolkit, so the check still depends on SPI, Trino plugins
and the plugin toolkit being versioned together and so remains
appropriate only for the plugins maintained within the project.