Change column name to avoid incorrect parse error message
Preparatory commit before `VALUE` is made a non-reserved keyword
For the purpose of supporting the JSON_OBJECT function.

This commit changes a colmn name from `value` to `v` in a test
which verifies error messages on parse exceptions.

If a query contains an identifier being a non-reserved token, and
later in te query text there's parse error, the error message points
to the correct location of the error, but it incorrectly returns the
non-reserved keyword as the erroneous token.