Enable Error Prone check: HidingField
This warning (now error) triggers when a subclass declares a field with
the same name as a non-private (i.e. visible) field in a superclass.
Usually this is a simple matter of redundancy, but there's one case
where the field in superclass is actually not useful - it is only used
in subclasses, and the assignment is trivial, so it's better declared in
them instead.