Remove grouped execution config options and session properties
This commit removes the feature of grouped execution. The code associated with
this feature is, however, mostly still present in the codebase. The commit only
removes config options and system session properties. The only additional
changes are for the purpose of successful compilation.
The dead code that is left in the project will be cleared in the following commits