Bigint hardcoded join code path
In the majority of cases the join is on a single bigint column. This commit
introduces a specific code path that will handle only single column bigint
joins. This way we can skip the logic behind value comparisons.

The new class holds bigint values in a long[] array. This provides far superior
value comparison performance. It comes, however, with a higher memory consumption.
That is why the limit of probe side positions is introduced, beyond which the old
implementation is chosen.