Fix Phoenix connector compilation in IntelliJ
IntelliJ doesn't seem to support multi-module projects when a module
being dependent on is built with shading.

As a workaround, add original dependency to compile-time classpath, but
exclude it from final build. This let's IntelliJ compile the project,
while still keeping the patched & shaded library as the runtime
dependency.

Alternatively we could move `trino-phoenix5-patched` module to a
separate repo. Not doing this in hope this is a temporary workaround
only.