SQL MERGE throw if assigning null to non-null column
This commit changes the two implementations of
MergeRowChangeProcessor to throw an exception if a
null value is assigned to a non-null column as a
result of an SQL MERGE operation.  An alternative
approach would have been to change the query plan
to insert such checks, but the chosen approach is
simpler.

This commit adds a BaseConnectorTest showing that SQL MERGE
prevents assignment of NULL to non-null target table columns.