Test varchar cast to date in predicate
The case where varchar is cast to date and used in a predicate is not
uncommon and it would be worth optimizing for. This involves, however,
some edge cases which are easy to forget about when working on such an
optimization (either within the engine, or on the connector level).