Analyze the context of table function invocation
Table sample or MATCH_RECOGNIZE should not be applied
to a polymorphic table function invocation.

Per SQL standard, relation alias is required for a polymorphic
table function with the GENERIC TABLE or static returned type,
and it is not allowed for table function with ONLY PASS THROUGH
returned type. We only enforce the latter rule.