Verify SQL MERGE access control checks
This commit adds a MERGE test in TestAccessControl that verifies
that in order execute a MERGE, you must have appropriate access
control rights.  Specifically, if there is a source table, you must
have SELECT on the table columns referenced; if the MERGE does
inserts, you must have INSERT on the target table; if the MERGE
does deletes you must have DELETE on the target table; and if
the MERGE does updates, you must have UPDATE on the target
table.