Transition task to FINISHED after noMoreOperators is set
Currently dynamic filters are delivered via DynamicFilterFetcher.
When a TaskStatus contains an updated dynamic filter version the
DynamicFilterFetcher is responsible for fetching the new dynamic
filters from a worker. However tasks were getting transitioned
to FINISHED before the LocalDynamicFilterConsumer#setPartitionCount
is set and dynamic filters updated. When task is transitioned
to FINISHED the TaskStatusFetcher no longer tries to update TaskStatus
so the coordinator is not able to learn that there is dynamic filters to
be fetched.