Fix improper expression aliasing during pushdown
EqualityInference was considering an expression of the form
cast(null AS ...) as an inference candidate, even though it
doens't satisfy the condition of not returning nulls on null
input.

As a result, unrelated expressions were being incorrectly equated
to each other. In particular, for the example in the test,
CAST(null AS varchar) was being considered equivalent to VARCHAR 'hello'
and then substituted in the expression:

    CAST(CAST(null AS varchar) AS bigint)

to produce the (invalid) expression:

    CAST(VARCHAR 'hello' AS bigint)