Avoid cleaning up metadata files in case of AWS Glue failure
In case of dealing with an `RuntimeException` while
commiting an Iceberg transaction, the Iceberg framework
will clean up the metadata file corresponding to the transaction.
In case of a metastore client timeout operation the
Iceberg library can therefore delete metadata files
which eventually get referenced from the configuration
of the table persisted on the metastore for the table
which leaves the table in a corrupt state.

Throw `CommitStateUnknownException` when dealing
with AWS Glue exceptions to ensure that the
table is not left in a corrupt state after the erroneous
completion of the DML operation.