Reset allowed-no-matching-node-period timeout on new nodes in cluster
We allow for situation when there are no nodes in the cluster which
match task requirement for timeout governed by
node-scheduler.allowed-no-matching-node-period config variable.
Previously if new node appeared in the cluster, but could not be used
because it did not have enough free resources at given moment timeout
counter was not reset.
This commit changes the behaviour. If new node appears and then
disappears again, the no-nodes timeout will be recounted again from
zero.