Traverse sources only once in RemoveEmptyUnionBranches
The `isEmpty` was used to first check whether there is work to do, and
then done once again to check which sources to exclude. This can be done
once per source only.

The added cost of building new sources and symbol mapping should not
matter, as it's a local operation, and `isEmpty` needs to traverse the
whole plan.