Clarify ConnectorMetadata listing when schema name not found
It's important to require that `listTables`, `listViews` and
`listMaterializedViews` do not fail when asked to list relations from a
schema that does not exist. This is because schema existence cannot be
ensured in a concurrent environment, i.e. a schema could have just been
deleted concurrently. Lack of such assumptions for these methods would
make predicate pushdown for tables like `system.jdbc.tables` impossible
to do properly.