Fix Hive test table name collision
In TestHiveTransactionalTable both testAcidUpdateSelectedValues and
testAcidUpdateSimple use the root test name test_update_simple, and
withTemporaryTable did not add a random suffix, so the the two tests
using that root name collided.  This PR changes the table name used
in the former test, and also changes withTemporaryTable to add a
random table name suffix.