Scale partitions to multiple writers locally
First figure out if there's a skewness across
writers in a node, then find the biggest partitions
in the skewed writers and scale them across writers
which are on the lower end i.e. written smallest
amount of data.

Scaling will only happen if the skewness is above 70%
and the partition to be scaled has written atleast
writerMinSize since last scale up.