Fix set-milestone failure when latest milestone is closed
There can be situation where the latest milestone is closed but there
are still some PRs for which this workflow is pending. The  API by
default only lists open milestones. This commit changes it to list all
milestones so that even if the target milestone is closed it can be
found and the PR added to it.

This is correct reimplementation of
9d66416c2d0e37fdfba84afd1845ad7a1f043aad.