Add most properties to Identity's hashCode and equals
The rationale is that, in addition to the user name, all of these can
affect access control decisions for an `Identity`, so they should be
considered distinct if any of those attributes differ.

The `StatementAnalyzer`, during the analysis stage, collects all table
references along with their corresponding access controls and
identities. At the end of the analysis, this is used to do the access
checks. They are collected in a `Map`, where the key is the pair of
access control and `Identity`, so that we deduplicate the necessary
checks. But for this deduplication to work properly, `Identity`
instances that result in different accesses should be considered
distinct.

Remove obsolete AccessControlUtil.FullIdentityEquality.
Now `Identity` itself has full identity.