Add requiredColumns field to TableFunctionAnalysis
The new field allows the table function to declare during
analysis which columns from the input tables are necessary to
execute the function.

The required columns can be then validated by the analyzer.
This declaration can be also used by the optimizer to prune
any input columns that are not used by the table function.