Adjust scheduler assignment queue for node
Add splits assignment adjustment based on node statistics. If the node is able to process all splits in the queue,
the amount of max pending splits to be assigned is doubled in the next iteration of the computeAssignments method.
Scaling down is done after interval from previous scaling up.
Queue adjustment is limited to be in range of newly created `min-pending-splits-per-task` and changed `max-pending-splits-per-task` to `max-adjusted-pending-splits-per-task`.