Fix incorrect results with empty projections in BigQuery
When the BigQuery connector noticed an empty projection or a query which
can be answered without adding any projections then it uses the
TableInfo in case of tables (not views or other table types) to find the
row count and generate pages instead of actually querying BigQuery.

The row count as shown in TableInfo for tables written using the
insertAll/streaming API is incorrect until the streaming buffer is
flushed and there's no mechanism to trigger an on-demand flush which can
lead to incorrect results.

To fix this we now execute a COUNT(*) query (similar to what was already
done for views) when we want to generate pages for empty projections.