Fix killing threads in Matcher
Before this change, killing threads in the Matcher was incorrectly implemented.
When a thread was killed, it still persisted in the threadsAtInstructions
structure. However, all other thread's data was cleared, so accessing the
thread caused failure.

The simplest solution would be to remove the thread from the threadsAtInstructions
structure while killing the thread. However, that would be a costly operation,
because there might be multiple entries for the thread across the structure.
Additionally, by removing the thread eagerly, we would miss the opportunity
to early kill other threads, equivalent to the removed thread.

The chosen solution is to delay killing threads until the end of the iteration
of the main loop of the matching algorithm. Instead of killed, the invalidated threads
are recorded as threadsToKill. The downside to this solution is higher memory
consumption during the iteration. The advantages include more efficient thread pruning
and easy maintenance of the threadsAtInstructions structure (it is reset after
the iteration).