Allow specifying exact desired split count
Some tables (such as store_returns, store_sales) change the number of
rows based on the number of splits generated. While it is unclear
whether this is expected behavior (as the same is observed in the
reference implementation from TPC) for now it was decided to allow
specifying the number of splits explicitly to allow generating
deterministic data sets.