Increase task.scale-writers.max-writer-count to core count
Set the value of default max writer count to the
number of processors and cap it to 32. We can do
this because preferred write partitioning is always
enabled for local exchange thus partitioned inserts
will never use this property. Hence, there is no
risk in terms of more numbers of physical writers
which can cause high resource utilization.