Obtain and refresh sink instance handle asynchronously
It is not always possible for exchange implementation to
provide an immediate return for calls to:
* `Exchange.instantiateSink `and
* `Exchange.updateSinkInstanceHandle`

depending on exchange implementation sometimes there needs to be a network
the communication involved or another lengthy process. To allow for such
exchange implementations and not block engine threads this PR changes
SPI so the two methods listed above now return
`CompletableFuture<ExchangeSinkInstanceHandle>`.