Avoid saving cache in the build-test-matrix CI job
This job is not supposed to download all dependencies, so it should not
be saving the local Maven repository in the GitHub Actions cache. Since
it can run quick than other jobs that do a full build, it can generate
an incomplete cache entry, preventing bigger cache entries from being
saved.