Fix reading parquet column with unused dictionary
A parquet file produced by Impala was found to have an empty dictionary
which is not used in the encoding of data pages in the column.
For such a case we cannot rely on ColumnChunkMetaData#hasDictionaryPage
as that checks for whether the data pages are also encoded using the dictionary.
This change removes usage of hasDictionaryPage to fix query failures
with such files.