Don't enforce top-level repo name in BaseCostBasedPlanTest
Engineers will typically check out the repo into a local folder named
the same way: `trino`. This, however, doesn't work when an engineer
needs to do a second checkout, or wants to make a backup of the repo for
whatever purposes. The build and tests should not assume any particular
name of the folder the repo is checked out into.