Remove redundant table metadata calls
This retrofits existing `ConnectorMetadata.getSchemaTableName`. The
method remains deprecated, because a new name, `getTableName` is better.

As a benefit of this change, the `ConnectorMetadata`'s
`getTableMetadata` and `getTableSchema` no longer need to support
synthetic table handles. This fixes a bug in connectors that support
advanced pushdown. For example, `HiveMetadata.getTableMetadata` called
for `HiveTableHandle` having `projectedColumns` should return column
metadata for the projected columns, but it did not do that.