Fix HiveMetadata.streamTableColumns not to include views
In the absence of explicit documentation, the use place
(`io.trino.metadata.MetadataManager#listTableColumns`) defines
`streamTableColumns` contract as covering tables only.  This commit also
supplements explicit documentation in `ConnectorMetadata`.

This relates to a logic introduced in `HiveMetadata.doGetTableMetadata`
in 6cb505b7676c13f9748da8f0af542b5ccb40e8ef commit. It was practical for
hive views and when view translation is not enabled, but was not correct
for trino views.