Create a fault-tolerant execution smoke test for `trino-redshift`
Opt for a smoke test to verify basic functionality of a CTAS statement
on the `trino-redshift` connector.
Running all the fault-tolerant execution tests on the connector can
occasionally lead to timeouts due to the random geographical positioning
of the Github CI runner in relation to the ephemeral AWS Redshift instance.