Approximate written bytes in Hive and Iceberg sorted writers
Hive and Iceberg use the written bytes value to determine when a new file
should be started. Without this value the `target-max-file-size` config
property is ignored.

While writing, the returned value is an approximation. However, an accurate
value is returned after the writter commits.

Doing this does reduce the sort window to one file rather than all the
data the PageSink receives. Broadening the sort window while also splitting
files into the right size can be done as follow up.