Lower KRB ticket lifetime in EnvMultinodeTlsKerberosDelegation
In 36b7553f219cb47b3abbcd452101ba0c3db6d0b4 the HadoopKerberos
environment's ticket lifetime was lowered to be able to test Keberos
constrained delegation support in the JDBC driver. The HadoopKerberos
environment is a base environment and multiple environments inherit from
it. Changing the Kerberos config should be done in dedicated
environments to make sure we don't indicentally affect other envs or
tests which depend on HadoopKerberos environment.