Cast char fields, if necessary, to varchar type in view translations
Bump coral to 2.2.14 to support Hive view translation cnsisting of a UNION
between a char and a varchar field.

Due to wrong coercion between `varchar` and `char` in Trino, as described in
trinodb/trino#9031
, a work-around needs to be applied in case of translating Hive views which
contain a UNION dealing with char and varchar types.
The work-around consists in the explicit cast of the field
having char type towards varchar type corresponding of the set operation
output type.