Reduce test flakiness
Once the table is created, attempt to get the stats a few times
while waiting for PostgreSQL to analyze the table and declare
success as soon as the stats are available.

Previously, every iteration of the test had a race condition
between creating the table and checking for stats, which led
to many of the iterations to fail.