Remove unnecessary logic from CachingJdbcClient
This logically reverts commit 7c8a926e34bf7d84a240310c4115ee33dd35ad75.
The class uses `EvictableCacheBuilder` and the
`io.trino.cache.EvictableCacheBuilder#cacheDisabled` considers the cache
to be disabled when TTL is 0 or cache size is 0, so no need to set cache
size to 0 if TTL is known to be 0.