Remove testReadMetadataWithRelationsConcurrentModifications test
This test covers an important aspect of Trino connectors -- metadata
listing should work also while tables/views/materialized views are being
created or dropped. It helped discover real problems, e.g.

- in Iceberg, that's why it was created
- in Kudu: https://github.com/trinodb/trino/issues/12974,
- in SQL Server https://github.com/trinodb/trino/blob/7c914aec6585e792b05d20cb7f169fa775503cb5/plugin/trino-sqlserver/src/test/java/io/trino/plugin/sqlserver/BaseSqlServerConnectorTest.java#L151-L152

However, this turned out to be hard to test. The existing test is pretty
complicated, takes a lot of time, has high timeout value and remains one
of the more flaky tests. The test author does not know how to make it
robust and perhaps does not want to remain as the author of the top
flaky test in the project.