Use explicit values for MaxResults in all Glue APIs
Most listing Glue APIs accept a MaxResults parameter which decides how
many objects are returned in a single API call. The default values are
not documented but observed behaviour is that the default values change
on some basis.

This commit adds explicit values for MaxResults in the APIs which
support it to the maximum possible values.

This possibly reduces the number of Glue calls when listing tables,
databases or functions in some cases.

This is similar to 52928c57749f25a1ca1c1c230add03843ae8d75f and
5d169c28463f34121a9b333bcde6a9fda947fc99.