Verify user-visible message in assertQueryFails
Change `assertQueryFails` so that it verifies the message that the
end-user would see. Before the change, `assertQueryFails` would accept
any TrinoException that has expected message somewhere in the causal
chain. However, causal chain may or may not be visible to the end user,
depending on the tooling used. This restores the behavior that used to
be before commit a05ed1980fee8318bdbc94d1c258ad0fd6a7df56.

Tests that want to verify causal message should do this explicitly.