Ensure complete TaskInfo/Status for FINISHED tasks
Commit ensures that for tasks which completed successfuly (task state
is set to FINISHED) we construct TaskInfo/TasksStatus with state set to
FINISHED only when TaskHolder transtions out of initial state when it is
not referencing SqlTaskExecution.

Otherwise it is possible that coordinator would get incomplete
information about the task, which is already marked as FINISHED. In such
case no longer ask for updates for updates. This may happen for very
short lived tasks.

One of the important things which would be missing in such a case is
dynamic filter information.