Do not use SkewedPartitionRebalancer for remote exchange in FTE
The rebalancing logic from SkewedPartitionRebalancer is already covered
by partitions splitting/merging in HashDistributionSplitAssigner.
Also the way remote SkewedPartitionRebalancer is configured does not
match FTE behaviour as there is assumption that data will be consumed by
N tasks where N is number of cluster nodes. This is not true for FTE.