Bump default value of max split count per task for FTE
With limit of 256 splits per task it was possible to end up with very
small tasks and possibly even reach limit of number of tasks per stage
(2^^15). The problem appeared when splits were very small.
Task sizing should be mostly done based on split size; max
number of splits should be treated as extra limit to be reached only in
rare situations. This PR bumps default value 8x to 2048.