Add missing check in PreAggregateCaseAggregations
In cfe7c74f1b4467e0777b6ae88af326c62ae0f644, a check allowing for
the result of the call to optimize an expression to be a non-constant
expression was lost:

    Object defaultValue = optimizeExpression(caseExpression.defaultValue().get(), context);
    if (defaultValue != null) {
      ...
    }

As a result, the optimization produced plans with references to invalid columns.

This change restores the check and cleans up the nested if/else structure to make it more readable.