Make max requests configurable in Azure native FS
Max requests where increased from 5 to 2 * number of CPUs in
203436fb285b749f3ce32a9f97a09eb51d0dbf25. With a very large number of
nodes in a cluster, this can cause throttling and result in failed
queries. Make it configurable allows to tune the value, if the default
one is too big.

This change also sets the max requests and max requests per host to the
same value, to limit the number of new configuration options.