Keep most recent task failure in failures history
We are keeping history of task failures which is appended to final
failure of query if we run out of retries. The last task failure which
caused the query to fail was not part of the history. It was root
exception but task Id was missing. With this PR we will keep it both as
root exception and as part of history.