Generate job annotations without GHA REST API
The separate annotate workflow was added to work around permission
issues in forked repositories. This is not needed when not using the GHA
REST API.

Adding annotations via the API was not working correctly
anyway, since it added a check for a commit, but one commit can have
multiple runs associated with it, and the check sometimes did not show
up in the main run.