Add flag to mark resource groups as disabled
Instead of setting hardConcurrencyLimit and maxQueuedQueries to 0 for
deleted groups, mark them explicitly as disabled. This approach allows
the removal of groups that were previously leaves. To prevent query
starvation, groups are only fully removed once they no longer have any
queued queries. If we don’t wait until the queue is empty, the parent
group (which has now become a leaf) could potentially starve the queued
queries, as queries assigned to it would be processed with higher
priority.