Fix S3InputStream's handling of large skips
When the skip(n) method is called the MAX_SKIP_BYTES check is skipped,
resulting in the call potentially blocking for a long time.

Instead of delegating to the underlying stream, set the nextReadPosition
value. This allows the next read to decide if it is best to keep the existing
s3 object stream or open a new one.

This behavior matches the implementations for Azure and GCS.