Refactor QueryDataDecoder to return Iterator
Iterable<List<Object>> was an initial sin, as the assumption was that you can iterate over data
multiple times, which isn't true anymore for spooled segments, which are downloaded and acknowledged
once the data is fully read.

Two new classes are introduced: SpooledSegmentIterator which is responsible for loading, acknowledging
spooled segments, closing of the input stream for the spooled segment. This encapsulates this logic
now in a single place and makes this logic reusable for other use cases. This iterator is lazy,
the segment is loaded on the first row read.